requirements.txt
asgiref==3.8.1
colorama==0.4.6
Django==5.0.7
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
drf-yasg==1.21.7
flake8==7.1.0
inflection==0.5.1
iniconfig==2.0.0
mccabe==0.7.0
packaging==24.1
pluggy==1.5.0
pycodestyle==2.12.0
pyflakes==3.2.0
PyJWT==2.8.0
pytest==8.3.2
pytest-django==4.8.0
pytz==2024.1
PyYAML==6.0.1
setuptools==71.1.0
sqlparse==0.5.1
tzdata==2024.1
uritemplate==4.1.1
